requirements.txt
langchain 
langchain-openai
langchain-core 
langchain-community
openai 
azure-cognitiveservices-speech 
scikit-learn
requests
python-dotenv